fix compile warnings [and wrong typecast

+++ src/sys/fileio/mprint.c
@@ -174,11 +174,11 @@ PetscErrorCode PetscVSNPrintf(char *str,size_t len,const char *format,size_t *fu
   }
   ierr = PetscFormatConvert(format,newformat);CHKERRQ(ierr);
 #if defined(PETSC_HAVE_VSNPRINTF_CHAR)
-  flen = (size_t) vsnprintf(str,len,newformat,(char*)Argp);
+  flen = vsnprintf(str,len,newformat,(char*)Argp);
 #elif defined(PETSC_HAVE_VSNPRINTF)
-  flen = (size_t) vsnprintf(str,len,newformat,Argp);
+  flen = vsnprintf(str,len,newformat,Argp);
 #elif defined(PETSC_HAVE__VSNPRINTF)
-  flen = (size_t) _vsnprintf(str,len,newformat,Argp);
+  flen = _vsnprintf(str,len,newformat,Argp);
 #else
 #error "vsnprintf not found"
 #endif
